+++ src/main/java/com/supermartijn642/trashcans/generators/TrashCansRecipeGenerator.java
@@ -3,7 +3,6 @@ package com.supermartijn642.trashcans.generators;
 import com.supermartijn642.core.generator.RecipeGenerator;
 import com.supermartijn642.core.generator.ResourceCache;
 import com.supermartijn642.trashcans.TrashCans;
-import net.minecraft.world.item.Items;
 import net.neoforged.neoforge.common.Tags;
 
 /**
@@ -21,24 +20,24 @@ public class TrashCansRecipeGenerator extends RecipeGenerator {
             .pattern("AAA")
             .pattern("BCB")
             .pattern("BBB")
-            .input('A', Tags.Items.STONE)
-            .input('B', Tags.Items.COBBLESTONE)
+            .input('A', Tags.Items.STONES)
+            .input('B', Tags.Items.COBBLESTONES)
             .input('C', Tags.Items.CHESTS_WOODEN)
             .unlockedBy(Tags.Items.CHESTS_WOODEN);
         this.shaped(TrashCans.liquid_trash_can)
             .pattern("AAA")
             .pattern("BCB")
             .pattern("BBB")
-            .input('A', Tags.Items.STONE)
-            .input('B', Tags.Items.COBBLESTONE)
-            .input('C', Items.BUCKET)
-            .unlockedBy(Items.BUCKET);
+            .input('A', Tags.Items.STONES)
+            .input('B', Tags.Items.COBBLESTONES)
+            .input('C', Tags.Items.BUCKETS_EMPTY)
+            .unlockedBy(Tags.Items.BUCKETS_EMPTY);
         this.shaped(TrashCans.energy_trash_can)
             .pattern("AAA")
             .pattern("BCB")
             .pattern("BBB")
-            .input('A', Tags.Items.STONE)
-            .input('B', Tags.Items.COBBLESTONE)
+            .input('A', Tags.Items.STONES)
+            .input('B', Tags.Items.COBBLESTONES)
             .input('C', Tags.Items.DUSTS_REDSTONE)
             .unlockedBy(Tags.Items.DUSTS_REDSTONE);
         this.shapeless(TrashCans.ultimate_trash_can)