+++ src/main/java/com/supermartijn642/trashcans/screen/UltimateTrashCanContainer.java
@@ -8,11 +8,6 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.ClickType;
 import net.minecraft.world.item.ItemStack;
-import net.neoforged.neoforge.items.IItemHandlerModifiable;
-import net.neoforged.neoforge.items.ItemStackHandler;
-import net.neoforged.neoforge.items.SlotItemHandler;
-
-import javax.annotation.Nonnull;
 
 /**
  * Created 7/11/2020 by SuperMartijn642
@@ -25,27 +20,47 @@ public class UltimateTrashCanContainer extends TrashCanContainer {
 
     @Override
     protected void addSlots(Player player, TrashCanBlockEntity entity){
-        this.addSlot(new SlotItemHandler(entity.ITEM_HANDLER, 0, 63, 25));
-        this.addSlot(new SlotItemHandler(entity.LIQUID_ITEM_HANDLER, 0, 93, 25));
-        this.addSlot(new SlotItemHandler(entity.ENERGY_ITEM_HANDLER, 0, 123, 25));
+        this.addSlot(new DummySlot(0, 63, 25){
+            @Override
+            public boolean mayPlace(ItemStack stack){
+                return UltimateTrashCanContainer.this.validateObjectOrClose() && UltimateTrashCanContainer.this.object.isRegularItemValid(stack);
+            }
+        });
+        this.addSlot(entity.LIQUID_ITEM_HANDLER.apply(0, 93, 25));
+        this.addSlot(entity.ENERGY_ITEM_HANDLER.apply(0, 123, 25));
 
         // item filter
-        for(int column = 0; column < 9; column++)
-            this.addSlot(new SlotItemHandler(this.itemFilterHandler(), column, 8 + column * 18, 64) {
+        for(int column = 0; column < 9; column++){
+            int slotIndex = column;
+            this.addSlot(new DummySlot(slotIndex, 8 + column * 18, 64) {
                 @Override
                 public boolean mayPickup(Player playerIn){
                     return false;
                 }
+
+                @Override
+                public ItemStack getItem(){
+                    return UltimateTrashCanContainer.this.validateObjectOrClose() ? UltimateTrashCanContainer.this.object.itemFilter.get(slotIndex) : ItemStack.EMPTY;
+                }
             });
+        }
 
         // liquid filter
-        for(int column = 0; column < 9; column++)
-            this.addSlot(new SlotItemHandler(this.liquidFilterHandler(), column, 8 + column * 18, 94) {
+        for(int column = 0; column < 9; column++){
+            int slotIndex = column;
+            this.addSlot(new DummySlot(slotIndex, 8 + column * 18, 94) {
                 @Override
                 public boolean mayPickup(Player playerIn){
                     return false;
                 }
+
+                @Override
+                public ItemStack getItem(){
+                    TrashCanBlockEntity entity = UltimateTrashCanContainer.this.object;
+                    return entity == null || entity.liquidFilter.get(slotIndex) == null ? ItemStack.EMPTY : entity.liquidFilter.get(slotIndex).getRepresentingItem();
+                }
             });
+        }
     }
 
     @Override
@@ -117,26 +132,5 @@ public class UltimateTrashCanContainer extends TrashCanContainer {
                 this.getSlot(index).set(ItemStack.EMPTY);
         }
         return ItemStack.EMPTY;
-    }
-
-    private IItemHandlerModifiable liquidFilterHandler(){
-        return new ItemStackHandler(9) {
-            @Nonnull
-            @Override
-            public ItemStack getStackInSlot(int slot){
-                TrashCanBlockEntity entity = UltimateTrashCanContainer.this.object;
-                return entity == null || entity.liquidFilter.get(slot) == null ? ItemStack.EMPTY : entity.liquidFilter.get(slot).getRepresentingItem();
-            }
-        };
-    }
-
-    private IItemHandlerModifiable itemFilterHandler(){
-        return new ItemStackHandler(9) {
-            @Nonnull
-            @Override
-            public ItemStack getStackInSlot(int slot){
-                return UltimateTrashCanContainer.this.validateObjectOrClose() ? UltimateTrashCanContainer.this.object.itemFilter.get(slot) : ItemStack.EMPTY;
-            }
-        };
     }
 }