+++ changelog.md
@@ -1,2 +1,2 @@
 ### Trash Cans 1.0.18
-- Initial release of Trash Cans for NeoForge
+- Initial release of Trash Cans for Minecraft 1.20.5 & 1.20.6